# back_end/controller/schema/login.go
package schema

type LoginInput struct {
	Email    string `validate:"required" json:"email"`
	Password string `json:"password"`
}

type LoginResponse struct {
	AccessToken string `json:"accessToken"`
}